+++ src/containerapp/azext_containerapp/tests/latest/test_containerapp_sessionpool.py
@@ -21,7 +21,7 @@ class ContainerappSessionPoolTests(ScenarioTest):
     @ResourceGroupPreparer()
     @SubnetPreparer(location=TEST_LOCATION, delegations='Microsoft.App/environments',
                     service_endpoints="Microsoft.Storage.Global")
-    def test_containerapp_sessionpool(self, resource_group, subnet_id, vnet_name, subnet_name):
+    def test_containerapp_sessionpool_e2e(self, resource_group, subnet_id, vnet_name, subnet_name):
         location = TEST_LOCATION
         self.cmd('configure --defaults location={}'.format(location))
 
@@ -42,6 +42,17 @@ class ContainerappSessionPoolTests(ScenarioTest):
             resource_group, sessionpool_name_python, 300), checks=[
             JMESPathCheck('name', sessionpool_name_python),
             JMESPathCheck('properties.containerType', "PythonLTS"),
+            JMESPathCheck('properties.provisioningState', "Succeeded"),
+            JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.lifecycleType', "Timed"),
+            JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.cooldownPeriodInSeconds', 300),
+        ])
+
+        # Create GpuBase SessionPool
+        sessionpool_name_gpu = self.create_random_name(prefix='spgpu', length=24)
+        self.cmd('containerapp sessionpool create -g {} -n {} --container-type GpuBase --cooldown-period {}'.format(
+            resource_group, sessionpool_name_gpu, 300), checks=[
+            JMESPathCheck('name', sessionpool_name_gpu),
+            JMESPathCheck('properties.containerType', "GpuBase"),
             JMESPathCheck('properties.provisioningState', "Succeeded"),
             JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.lifecycleType', "Timed"),
             JMESPathCheck('properties.dynamicPoolConfiguration.lifecycleConfiguration.cooldownPeriodInSeconds', 300),
@@ -84,7 +95,7 @@ class ContainerappSessionPoolTests(ScenarioTest):
 
         # List Session Pools
         sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()
-        self.assertTrue(len(sessionpool_list) == 2)
+        self.assertTrue(len(sessionpool_list) == 3)
 
         # Update Session Pool
         max_concurrent_session = 12
@@ -132,6 +143,7 @@ class ContainerappSessionPoolTests(ScenarioTest):
 
         self.cmd('containerapp sessionpool delete -g {} -n {} --yes'.format(resource_group, sessionpool_name_python))
         self.cmd('containerapp sessionpool delete -g {} -n {} --yes'.format(resource_group, sessionpool_name_custom))
+        self.cmd('containerapp sessionpool delete -g {} -n {} --yes'.format(resource_group, sessionpool_name_gpu))
 
         # List Session Pools
         sessionpool_list = self.cmd("containerapp sessionpool list -g {}".format(resource_group)).get_output_in_json()