+++ src/containerapp/HISTORY.rst
@@ -14,6 +14,7 @@ upcoming
 * 'az containerapp env premium-ingress': Deprecate `--min-replicas` and `--max-replicas` parameters, use workload profile scale instead.
 * 'az containerapp sessionpool create/update': Support `--probe-yaml`
 * 'az containerapp session stop': Support stop session for custom container sessions
+* 'az containerapp sessionpool create': Add `GpuBase` container type
 
 1.2.0b3
 ++++++